+++ include/quad_gap/config/QuadGapConfig.h
@@ -27,10 +27,10 @@ namespace quad_gap
             struct Robot 
             {
                 float r_inscr = 0.18;
-                double length = 0.7; /**< Robot length */
-                double width = 0.3; /**< Robot width */
-                double avg_lin_speed = 0.2; /**< Average linear speed */
-                double avg_rot_speed = 0.5; /**< Average rotational speed */
+                float length = 0.7; /**< Robot length */
+                float width = 0.3; /**< Robot width */
+                float avg_lin_speed = 0.2; /**< Average linear speed */
+                float avg_rot_speed = 0.5; /**< Average rotational speed */
                 int shape_id = 1; /**< Robot shape ID, 1 for rectangle,
                                     2 for circle, 3 for triangle */
                 bool use_geo_storage = false; /**< Use precomputed robot geometry storage */
@@ -38,9 +38,9 @@ namespace quad_gap
 
             struct Goal 
             {
-                double xy_global_goal_tolerance = 0.2;
-                double yaw_global_goal_tolerance = 0.1;    
-                double xy_waypoint_tolerance = 0.1;
+                float xy_global_goal_tolerance = 0.2;
+                float yaw_global_goal_tolerance = 0.1;    
+                float xy_waypoint_tolerance = 0.1;
             } goal;
 
             /**
@@ -67,23 +67,23 @@ namespace quad_gap
                 int halt_size = 5;           
                 bool robot_path_orient_linear_decay = true; /**< Enable linear decay of robot path orientation */
                 bool virtual_path_decay_enable = true; /**< Enable virtual path decay */    
-                double decay_factor = 0.0; /**< Decay factor for virtual path decay */
+                float decay_factor = 0.0; /**< Decay factor for virtual path decay */
                 bool use_bezier = true; /**< Use Bezier curve for trajectory generation */
             } planning;
 
             struct ControlParams 
             {
-                double k_drive_x = 3.5;
-                double k_drive_y = 3.5;
-                double k_turn = 0.5;
-                double v_ang_const = 0.0;
-                double v_lin_x_const = 0.0;
-                double v_lin_y_const = 0.0;
+                float k_drive_x = 3.5;
+                float k_drive_y = 3.5;
+                float k_turn = 0.5;
+                float v_ang_const = 0.0;
+                float v_lin_x_const = 0.0;
+                float v_lin_y_const = 0.0;
                 int ctrl_ahead_pose = 2;
-                double vx_absmax = 0.5;
-                double vy_absmax = 0.5;
-                double ang_absmax = 0.2;       
-                double speed_factor = 2.0;
+                float vx_absmax = 0.5;
+                float vy_absmax = 0.5;
+                float ang_absmax = 0.2;       
+                float speed_factor = 2.0;
             } control;
             
             struct ManualControl 
@@ -97,36 +97,36 @@ namespace quad_gap
 
             struct GapManipulation 
             {
-                double sigma = 1.0;
-                double reduction_threshold = M_PI / 2;
-                double reduction_target = M_PI / 4;
+                float sigma = 1.0;
+                float reduction_threshold = M_PI / 2;
+                float reduction_target = M_PI / 4;
                 int max_idx_diff = 256;
                 bool radial_extend = true;
                 bool radial_convert = true;
-                double rot_ratio = 1.5;
+                float rot_ratio = 1.5;
             } gap_manip;
 
             struct ProjectionParam 
             {
-                double k_po = 0.8;
-                double r_min = 0.5;
-                double r_norm = 0.75;
-                double r_norm_offset = 0.5;
-                double k_po_turn = 1;
+                float k_po = 0.8;
+                float r_min = 0.5;
+                float r_norm = 0.75;
+                float r_norm_offset = 0.5;
+                float k_po_turn = 1;
             } projection;
 
             struct Trajectory 
             {
-                double integrate_maxt = 50;
-                double integrate_stept = 1e-2;
-                double rmax = 0.3;
-                double cobs = -1;
-                double w = 3;
-                double inf_ratio = 1.2;
-                double terminal_weight = 10;
-                double robot_geo_scale = 1;
+                float integrate_maxt = 50;
+                float integrate_stept = 1e-2;
+                float rmax = 0.3;
+                float cobs = -1;
+                float w = 3;
+                float inf_ratio = 1.2;
+                float terminal_weight = 10;
+                float robot_geo_scale = 1;
                 bool bezier_interp = true;
-                double bezier_unit_time = 0.1;             
+                float bezier_unit_time = 0.1;             
             } traj;
